Refactor methods out of Access facade class
The Access class used to perform lookups of EDD entity objects had collected
some other methods that did not operate as lookup methods, but instead used
other implicit lookup methods. In this commit, those methods are hoisted out of
the class and into their own classes.

One set of methods is moved into a new class hierarchy, for providing accessor
functions to a Handsontable ColumnSetting object. These accessor factory
classes make use of an internal Access object to perform lookups.

Another set of changed methods takes an array of records, Line or Assay, and
returns a merged record. These methods were taking a JQuery object of form
fields containing values corresponding to the records to merge. Rather than use
an error-prone translation between records to JQuery/Elements and back to
records, the methods are changed to use just the array of records. Client code
is simplified by splitting the operations of building the array of records and
generating the JQuery object for those records.