Refactor Access object to its own module
The Access object is used as a facade to simplify reading information out of
the EDDData interface. It's used by the Config module to define how tables are
built, but it makes more sense to have Access as its own entity so that other
pieces of code that deal with EDDData object access, but not necessarily
dealing with tables, can still leverage the facade interface.