Split table view into assays and measurements
Previous version of the table view attempted to merge together a view of assays
and measurements, and as a result could be confusing. It was unclear where to
select different items, the row highlighting was inconsistent between the assay
side and the measurement side, etc. Instead, add controls to view a table of
either assays, or measurements. Simpler table view, easier to understand,
similar functionality.